Moved basic NSColor and font definitions to ConfUIFoundation framework
