+++ ConfUIFoundation/ConfUIFoundation.h
@@ -1,0 +1,19 @@
+//
+//  ConfUIFoundation.h
+//  ConfUIFoundation
+//
+//  Created by Guilherme Rambo on 03/06/20.
+//  Copyright © 2020 Guilherme Rambo. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for ConfUIFoundation.
+FOUNDATION_EXPORT double ConfUIFoundationVersionNumber;
+
+//! Project version string for ConfUIFoundation.
+FOUNDATION_EXPORT const unsigned char ConfUIFoundationVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <ConfUIFoundation/PublicHeader.h>
+
+#import <ConfUIFoundation/NSColor+Accent.h>
