+++ ConfUIFoundation/Source/Definitions/Fonts.swift
@@ -1,6 +1,6 @@
 //
 //  Fonts.swift
-//  WWDC
+//  ConfUIFoundation
 //
 //  Created by Guilherme Rambo on 23/05/17.
 //  Copyright © 2017 Guilherme Rambo. All rights reserved.
@@ -8,7 +8,20 @@
 
 import Cocoa
 
-extension NSAttributedString {
+public extension NSFont {
+
+    static func wwdcRoundedSystemFont(ofSize size: CGFloat, weight: NSFont.Weight = .regular) -> NSFont {
+        guard let desc = NSFont.systemFont(ofSize: size, weight: weight).fontDescriptor.withDesign(.rounded) else {
+            assertionFailure("Failed to get font descriptor")
+            return NSFont.systemFont(ofSize: size, weight: weight)
+        }
+
+        return NSFont(descriptor: desc, size: size) ?? NSFont.systemFont(ofSize: size, weight: weight)
+    }
+
+}
+
+public extension NSAttributedString {
 
     static func attributedBoldTitle(with string: String) -> NSAttributedString {
         let attrs: [NSAttributedString.Key: Any] = [
@@ -42,6 +55,6 @@ extension NSAttributedString {
 
 }
 
-extension NSFont {
+public extension NSFont {
     static let boldTitleFont = NSFont.wwdcRoundedSystemFont(ofSize: 24, weight: .semibold)
 }
