+++ ConfUIFoundation/Source/Headers/NSColor+Accent.h
@@ -1,6 +1,6 @@
 //
 //  NSColor+Accent.h
-//  WWDC
+//  ConfUIFoundation
 //
 //  Created by Guilherme Rambo on 29/05/20.
 //  Copyright © 2020 Guilherme Rambo. All rights reserved.
